reduced even more code and get finally rid of the last two external variables.
git-svn-id: https://clonekeenplus.svn.sourceforge.net/svnroot/clonekeenplus/cgenius/trunk@1215 4df4b0f3-56ce-47cb-b001-ed939b7d65a6

+++ src/dialog/CDialog.h
@@ -59,7 +59,7 @@ public:
 	int m_int;
 	int m_min;
 	int m_max;
-	int m_length;
+	size_t m_length;
 	char m_inputmode;
 	std::string m_name;
 	std::vector<CDlgObject*> m_dlgobject;
